Intuitionize bits from df-bits to bitsfzo (#5079)
* copy 5ndvds3 and 5ndvds6 from set.mm to iset.mm

* Add bits syntax to iset.mm

This is the syntax and df-bits .  Copied without change from set.mm.

* copy bitsfval from set.mm to iset.mm

* copy bitsval from set.mm to iset.mm

* copy bitsval2 from set.mm to iset.mm

* copy bitsss from set.mm to iset.mm

* copy bitsf from set.mm to iset.mm

* copy bits0 from set.mm to iset.mm

* Add bits0e to iset.mm

Stated as in set.mm.  The proof has a small rearrangement of the logic
but is otherwise the same as the set.mm proof.

* copy bits0o from set.mm to iset.mm

* Add bitsp1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with the usual
changes for division and floor theorems.

* Add bitsp1e to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a small change
for differences in division theorems.

* Add bitsp1o to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes
for differences in floor and division theorems.

* Move integer supremum functions earlier in iset.mm

These are zsupcllemstep , zsupcllemex , zsupcl , zssinfcl ,
infssuzex , infssuzledc , infssuzcldc , suprzubdc , nninfdcex ,
zsupssdc , and suprzcl2dc .

* Add qdcle to iset.mm

The proof is direct from qdclt .

* Add nn0maxcl to iset.mm

Similar to zmaxcl and other such theorems but for NN0 .  The proof is
not hard from zmaxcl .

* Add bitsfzo to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with various changes
for differences between set.mm and iset.mm.

Includes lemma bitsfzolem (also stated as in set.mm and with a proof
based on the set.mm proof).

* Fix bits0o comment in set.mm and iset.mm

It had said zero where one is correct.

* Revise bitsfzo comment

The previous wording was strange since by default bits are either
0 or 1, and the inferred meaning of 'bits' as 'numbers m such that
the bit at 2^m is 1' is not noted in df-bits

Referring to bits "at positions" is intended to be similar to usages
such as "` M + 1 ` -th bit" or "zeroth bit" which we use elsewhere.