intuitionize the rest of the bits section (bitsmod through bitsinvp1) (#5088)
* Add bitsmod to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof, with
intuitionization at various steps.

* Add bitsfi to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes for
decidability and to use fzofig .

* Update eulerpart entry in mmil.html

We now have bits but also note the use of supp in the set.mm proof.

* Add bitscmp to iset.mm

Stated as in set.mm.  The proof needs intuitionizing for several of the
usual reasons but is otherwise the set.mm proof.

* copy 0bits from set.mm to iset.mm

* copy m1bits from set.mm to iset.mm

* Add bitsdc to iset.mm

This is extracted from a portion of the bitsfi proof.

* Add bitsinv1lem to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in
various places to show propositions are decidable or to adjust for
differences in modulus and division theorems.

* Rename dfss2 to ssalel in iset.mm

This is because dfss2 is a different theorem in set.mm.  In set.mm, this
is df-ss but we're not (at least yet) changing the definition in iset.mm
so we'll use a new name for this theorem.

* Add dfss2 to iset.mm

Another name for ~ df-ss which is more consistent with the naming
in the Metamath Proof Explorer.

* Add bitsinv1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes
for differences in finite set and modulus theorems and for showing
propositions are decidable.

* Add ficardon to iset.mm

This might be as close as we can get to ficardom or might not, but we
can at least prove this much.

* Add ackbij1 to mmil.html

Also, expand ackbijnn entry to say a bit more about what parts of this
are (more or less) proved and how the rest might be proved.

* Add bitsinv2 to mmil.html

* add bitsf1ocnv to mmil.html

* add bitsf1o to mmil.html

* add bitsf1 to mmil.html

* add 2ebits to mmil.html

* add bitsinv to mmil.html

* add bitsinvp1 to mmil.html