Add prdsval to iset.mm
Stated as in set.mm.  The proof needs a lot of changes for set existence
and similar differences but is largely the set.mm proof.

Also adds $d f g K $. which should not cause a problem given the usages
of this theorem, as far as I noticed.  Since f and g are bound in the
hypotheses, it should be possible to remove this condition if needed.