Intuitionize structure product from imasvalstrd to prdsmulr (#5096)
* Add imasvalstrd to iset.mm

This is imasvalstr from set.mm with set existence conditions added.  The
proof is similar to the set.mm proof.

* Add theorems for oc slot.

This is ocndx , ocid , basendxnocndx and plendxnocndx .
Copied without change from set.mm (except for small changes to
comments).

The iset-discouraged entries for basendx and ocndx are also found in
set.mm.

* Copy homndx from set.mm to iset.mm

Update iset-discouraged .

* copy ccondx from set.mm to iset.mm

Update iset-discouraged

* Add prdsvalstrd to iset.mm

This is prdsvalstr from set.mm with set existence conditions added.  The
proof is similar to the set.mm proof.

The new usages of ccondx and homndx in iset-discouraged are also usages
in set.mm and found in discouraged .

* Add ovssunirng to iset.mm

This is ovssunirn from set.mm with added set existence conditions.  The
proof is similar to the set.mm proof.

* Add prdsvallem to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but many
parts need changes for differences in set existence theorems.

* Update df-prds in iset.mm

This is a verbatim copy of the current df-prds (updated
18-Aug-2024) from set.mm.  Adjust reldmprds and prdsex proofs
accordingly.

* Add prdsval to iset.mm

Stated as in set.mm.  The proof needs a lot of changes for set existence
and similar differences but is largely the set.mm proof.

Also adds $d f g K $. which should not cause a problem given the usages
of this theorem, as far as I noticed.  Since f and g are bound in the
hypotheses, it should be possible to remove this condition if needed.

* Revise strslfv3 in iset.mm

Put more hypotheses in deduction form.  Because this theorem is unused
in iset.mm and does not occur (under this name) in set.mm, this change
is possible.

* Add prdsbaslemss to iset.mm

This is like prdsbaslem from set.mm in the sense that it is a lemma for
theorems like prdsbas but many of the details are different.

Also add prdssca and prdsbas which are stated as in set.mm.  Their
proofs are similar to the set.mm proofs but adapted to use
prdsbaslemss instead of prdsbaslem .

* Add prdsplusg to iset.mm

Stated as in set.mm.  The proof is generally similar to the set.mm proof
but greatly changed for differences between set.mm and iset.mm.

* Add prdsmulr to iset.mm

Stated as in set.mm.  The proof needs various changes for differences in
iset.mm but is generally similar to the set.mm proof.

* Expand prdsvsca entry in mmil.html

Say a little bit about how it, and similar theorems, could be proved.