add 2omap to iset.mm and "The size of sets" section to mmil.html (#5101)
* Add 2omap to iset.mm

That ( 2o ^m A ) can be thought of as decidable subsets of A is by now a
familiar idea, but here we make it precise with something we can prove.

* Add $j usage in iset.mm for f1oen3g and friends

This matches some of the $j usage directives in set.mm.  It is for
f1oen3g , f1oen2g , and f1dom2g .

* Add 2omapen to iset.mm

Corollary of 2omap

* Move A Theorem Sampler earlier in mmil.html

Since the purpose is to give a starting point, it seems like it is
better to include it early, before we start diving into topics which
will likely seem obscure especially at first.  This better matches its
position within mmset.html.

* Add "The size of sets" section to mmil.html

Loosely based on similar material in AczelRathjen and BauerHanson
but as usual with an eye towards iset.mm notation and mentioning
iset.mm theorems.