Add psr1clfi to iset.mm
This is psr1cl from set.mm with the additional condition that the index
set must be finite.  The proof is the set.mm proof with some small
additions for set existence and a much larger addition for decidability
(this is also the part which requires a finite index set).