Intuitionize psrgrp and related theorems (#5116)
* Add psrlinv to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with more steps
in a few places for set existence.

* Add psrgrp to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a few more
steps for set existence.

* copy psr0 from set.mm to iset.mm

* copy psrneg from set.mm to iset.mm

* Add psrlmod to mmil.html

* Add psr1clfi to iset.mm

This is psr1cl from set.mm with the additional condition that the index
set must be finite.  The proof is the set.mm proof with some small
additions for set existence and a much larger addition for decidability
(this is also the part which requires a finite index set).

* add psrlidm to mmil.html

Also revise entries for psrmulr , psrmulfval , and psrmulval to
more fully discuss the multiplication slot for mPwSer .

Revise psrbagcon entry to discuss what would be needed a bit more.

* Add psrridm to mmil.html

* Add psrass1 to mmil.html

* add psrdi to mmil.html

* add psrdir to mmil.html

* Add psrass1 to mmil.html

* Expand mmil.html text on lgsqr lemmas

* add psrcom to mmil.html

* add psrass23 to mmil.html

* add psrring to mmil.html