Add funop etc to iset.mm
Includes funiun .  Stated as in set.mm and with a proof based on the
set.mm proof (with some small changes)

Includes funopsn .  Stated as in set.mm.  The proof is a new one based
on funopg and other existing theorems.

Copy funop and funopdmsn from set.mm to iset.mm