add mPoly to iset.mm (#5130)
* add mPoly to iset.mm

The definition is intended to accomplish what the set.mm definition does
but is different as noted in the comment.

* Add reldmmpl to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof.

* Add mplvalcoe to iset.mm

This is mplval from set.mm with added set existence conditions and
which changes the expansion of the value to reflect ~ df-mplcoe .
The proof is similar to the set.mm proof.

* Add mplbascoe to iset.mm

This is mplbas from set.mm with set existence conditions added and
the conclusion modified to reflect df-mplcoe .  The proof is similar
to the set.mm proof.

* Add mplelbascoe to iset.mm

This is mplelbas from set.mm with added set existence conditions and
adjustments for differences in the mPoly definition.  The proof is
similar to the set.mm proof.

* Add df-mvr and mvrcl to mmil.html

* add mvrf2 to mmil.html

* Add fnmpl to iset.mm

This is similar to other existence/domain theorems for similar
notations.

* Add mplrcl to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof.

* add mplelsfi to mmil.html

* Add mplval2g to iset.mm

This is mplval2 from set.mm with set existence conditions added.  The
proof is similar to the set.mm proof.

* Add mplbasss to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof but is
longer due to having more set existence machinery.

* copy mplelf from set.mm to iset.mm

* Add mplsubg lemmas to mmil.html

This is mplsubglem , mpllsslem , and mplsubglem2 .  However we
end up proving mplsubg , it is unlikely to be via these lemmas.

* Add psrbagfi to iset.mm

Turns out that if we are willing to assume the index set is finite in
theorems that use mPwSer , that removes a whole set of hassles around
finite support.

* Add psrelbasfi to iset.mm

Although this is a simple consequence of psrelbas and psrbagfi ,
it should declutter proofs using it significantly.

* Add mplsubgfi to iset.mm

This is mplsubg from set.mm but the index set needs to be finite.
The proof is very loosely based on the set.mm proof but many things are
changed, including for the differences in the mPoly definition.

Includes lemmas mplsubgfilemm , mplsubgfilemcl , and mplsubgfileminv .

* add mpllss to mmil.html

* add mplsubrglem , mplsubrg to mmil.html

* Add mpl0fi to iset.mm

This is mpl0 from set.mm where the index set needs to be finite.  The
proof is the set.mm proof with adjustments at various steps but no
fundamental changes.

* Add mplplusgg to iset.mm

This is mplplusg from set.mm with set existence conditions added.  The
proof is the set.mm proof with additional steps for the set existence.

* Add mplmulr to mmil.html

* Add mpladd to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with additional
steps for set existence.

* Add mplnegfi to iset.mm

This is mplneg from set.mm where the index set needs to be finite.  The
proof is basically the set.mm proof but needs changes in several places.

* add mplmul to mmil.html

* add mpl1 to mmil.html

* add mplvsca to mmil.html

* add mplvscaval to mmil.html

* Add mplgrpfi to iset.mm

This is mplgrp but needs a finite index set.  The proof is the set.mm
proof with some small changes.

* add mpllmod to mmil.html

* Add mplring to mmil.html