Add 2strstrndx to iset.mm
This is like 2strstrg in iset.mm but adapted to current conventions
around extensible structure indexes.  It is the same as 2strstr in
set.mm but with set existence conditions added.  The proof is slightly
modified from the 2strstrg proof.

Mark 2strstrg as discouraged because it hardcodes the index one.