Intuitionize the last element of a word and the concatenation of two words (#5141)
* Add lastS to iset.mm

This is the syntax, some additional text in a section header, and df-lsw

All copied without change from set.mm.

* Add lswwrd to iset.mm

This is lsw from set.mm but only defined for words.  The proof is
similar to the set.mm proof with a few more steps for set existence.

* Add nnnle0 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with several small
changes.

* Add lsw0 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with several small
changes.

* copy lsw0g from set.mm to iset.mm

* Add lsw1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a small change.

* Add lswcl to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with one small
change.

* copy lswlgt0cl from set.mm to iset.mm

* Add ++ to iset.mm

This is the syntax , df-concat , and some text in the section header.

Copied without change from set.mm.

* add ccatfn to mmil.html

* Add ccatfvalfi to iset.mm

This is ccatfval with the added condition that the operands are finite.
The proof is similar to the set.mm proof although several more steps are
needed to show set existence.

* Add ccatcl to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with several
additional steps, for example for decidability and closure.

* Add ccatlen to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in some
places including additional steps especially for set existence.

* Add ccatclab to iset.mm

This is ccatcl but where the two words are over different sets.  The
proof is via sswrd and ccatcl .

* Add ccat0 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a few
additional steps for finiteness.

* copy syldbl2 from set.mm to iset.mm

* copy fzo theorems from set.mm to iset.mm

This is elfzoextl , elfzoext and elincfzoext

* Add ccatval1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add ccatval2 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with several small
changes.

* copy fzo0addel and fzo0addelr from set.mm to iset.mm

* copy ccatval3 from set.mm to iset.mm

* copy elfzelfzccat from set.mm to iset.mm

* Add ccatvalfn to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof, although
there are some changes and rearrangements.

* Add ifeqdadc to iset.mm

This is ifeqda from set.mm with a decidability condition added.  The
proof is the set.mm proof with small changes for decidability.

* Add ccatsymb to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in
several steps.

* copy ccatfv0 from set.mm to iset.mm

* Add ccatval1lsw to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with lsw changed to
lswwrd .

* copy ccatval21sw from set.mm to iset.mm

* copy ccatlid from set.mm to iset.mm

* copy ccatrid from set.mm to iset.mm

* copy ccatass from set.mm to iset.mm

* copy ccatrn from set.mm to iset.mm

* copy ccatidid from set.mm to iset.mm

* Add lswccatn0lsw to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small
changes.

* copy lswccat0lsw from set.mm to iset.mm

* copy 3bior1fd from set.mm to iset.mm

* copy 3bior1fand from set.mm to iset.mm

* copy 3bior2fd from set.mm to iset.mm

* copy 3biant1d from set.mm to iset.mm

* copy intn3an1d from set.mm to iset.mm

* copy intn3an2d from set.mm to iset.mm

* copy intn3an3d from set.mm to iset.mm

* add mptfi to mmil.html

* Add konigsberg and ccatalpha to mmil.html

* add ccatrcl1 to mmil.html

* Fix comments to say nonempty word

This is to replace "not empty word" in set.mm and iset.mm.