Say Axiom of Collection in comments.
This is the most straightforward thing to do in iset.mm in the sense
that the axiom is called ax-coll and I guess in principle would be
stronger than zfrep6 although currently we only use ax-coll to prove
zfrep6 (and a handful of theorems not used further).  There isn't
anything going on differently than in the set.mm comments which say
"Axiom of Replacement".