Add singleton word constructor to iset.mm (#5142)
* Add singleton word constructor to iset.mm

This is the syntax , df-s1 and some text in the section header.
Copied without change from set.mm.

* add ids1 to mmil.html

* copy s1val from set.mm to iset.mm

* copy s1rn from set.mm to iset.mm

* copy s1eq from set.mm to iset.mm

* copy s1eqd from set.mm to iset.mm

* copy s1cl from set.mm to iset.mm

* copy s1cld from set.mm to iset.mm

* add s1prc to iset.mm

Stated as in set.mm.  The proof is via fvprc , df-s1 and s1val

* add s1cli to mmil.html

* Add s1leng to iset.mm

This is s1len from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof.

* Add s1nz to mmil.html

* Add s1dmg to iset.mm

The statement of the theorem and proof are taken from s1dmALT in set.mm.
The comment is taken from s1dm in set.mm.  Add s1dm and s1dmALT to
mmil.html.

* copy s1fv from set.mm to iset.mm

* Add lsws1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add eqs1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes
in various places.

* copy wrdl1exs1 from set.mm to iset.mm

* Add wrdl1s1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add s111 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* copy ccatws1cl from set.mm to iset.mm

* add ccatws1clv to mmil.html

* copy ccat2s1cl from set.mm to iset.mm

* add ccats1alpha to mmil.html

* Add ccatws1leng to iset.mm

This is ccatws1len from set.mm with a set existence condition on X .
The proof is similar to the set.mm proof.  Note that X does not need to
be an element of V .

* Add ccatws1lenp1bg to iset.mm

This is ccatws1lenp1b from set.mm with a set existence condition added.
The proof is the set.mm proof with small changes.

* add wrdlenccats1lenm1 to mmil.html

* add ccat2s1len to mmil.html

* copy ccatw2s1cl from set.mm to iset.mm

* add ccatw2s1len to mmil.html

* Add ccats1val1g to iset.mm

This is ccats1val1 from set.mm with a set existence condition added.
The proof is the set.mm proof with small changes.

* add ccats1val2 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add ccat1st1st to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes to
various steps.

* add ccat2s1p1 , ccat2s1p2 to mmil.html

* add ccatw2s1ass to mmil.html

* add ccatws1n0 to mmil.html

* copy ccatws1ls from set.mm to iset.mm

* Add lswccats1 to iset.mm

Stated as in set.mm.  The proof is via ccatws1ls .

* copy lswccats1fst from set.mm to iset.mm

* add ccatw2s1p1 to mmil.html

* Add ccatw2s1p2 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* add ccat2s1fvw to mmil.html

* add ccat2s1fst to mmil.html