Intuitionize graphs from snstrvtxval to edg0iedg0 (#5144)
* add funsndifnop to mmil.html

* add funsneqopb to mmil.html

* add snstrvtxval to mmil.html

* add snstriedgval to mmil.html

* Add vtxval0 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with one small
change.

* Add iedgval0 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add vtxvalsnop to mmil.html

* add iedgvalsnop to mmil.html

* add vtxval3sn to mmil.html

* add iedgval3sn to mmil.html

* copy vtxvalprc from set.mm to iset.mm

* copy iedgvalprc from set.mm to iset.mm

* Add Edg to iset.mm

This is the syntax and df-edg .  Copied from set.mm with the only change
being to a comment (to remove a reference to a theorem we don't have
yet).

* add edgval to mmil.html

* add iedgedg to mmil.html

* Add edgopval to iset.mm

Stated as in set.mm.  The proof is directly from df-edg rather than
via edgval

* copy edgov from set.mm to iset.mm

* Add edgstruct to iset.mm

Stated as in set.mm.  The proof needs to do a lot more than the set.mm
proof, partly because it includes a number of steps from edgval and
partly for other reasons including set existence.

* add edgiedgb to mmil.html

* Add edg0iedg0 to mmil.html