Intuitionize the substring of a word (#5145)
* Add substr to iset.mm

This is the syntax, df-substr , and some text in the section header.
Copied without change from set.mm.

* add nsyl5 to iset.mm

Stated as in set.mm.  The proof is via con3d .

* add nssdmovg to mmil.html

* reword mpondm0 in mmil.html

* add swrdnznd to mmil.html

* Add fzowrddc to iset.mm

This will be needed for many of the things we want to do
with df-substr .  The proof is via ssfzo12bi .

* Add swrdval to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes.

* Add swrd00g to iset.mm

This is swrd00 from set.mm with additional conditions.  The proof is
adapted from a portion of the set.mm proof.

* Add swrdclg to iset.mm

This is swrdcl from set.mm with additional conditions.  The proof is
taken from a portion of the set.mm proof with some changes.

* copy swrdval2 from set.mm to iset.mm

* Add swrdlen to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* add swrdfv to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* copy swrdfv0 from set.mm to iset.mm

* Add swrdf to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes needed
to use swrdclg instead of swrdcl .

* copy swrdvalfn from set.mm to iset.mm

* copy swrdrn from set.mm to iset.mm

* copy fzo0n from set.mm to iset.mm

* copy swrdlend from set.mm to iset.mm

* add jaoi2 , jaoi3 to mmil.html

* Add swrdnd to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in a
variety of places and some rearrangement of the logic.

* add swrdnnn0nd to mmil.html

* add swrdnd0 to mmil.html

* Add swrd0g to iset.mm

This is swrd0 from set.mm with additional conditions added.  The proof
is taken from a portion of the set.mm proof with small changes.

* copy swrdrlen from set.mm to iset.mm

* copy swrdlen2 from set.mm to iset.mm

* copy swrdfv2 from set.mm to iset.mm

* Add swrdwrdsymbg to iset.mm

This is swrdwrdsymb from set.mm with additional conditions added.  The
proof is based on a portion of the set.mm proof with some changes.

* copy swrdsb0eq from set.mm to iset.mm

* Add swrdsbslen to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes and some additional steps to show decidability.

* add swrdspsleq to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes and additional steps to show decidability.

* Add swrds1 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes to use
swrdclg in place of swrdcl .

* Add swrdlsw to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add ccatswrd to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add swrdccat2 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a few extra
steps to use swrdclg instead of swrdcl .