Add pfxwrdsymbg to iset.mm
This is pfxwrdsymb from set.mm with an additional condition.  The proof
is via pfxval and swrdwrdsymbg in the L <_ ( # ` S ) case (this part is
similar to the set.mm proof), but the additional length of the proof
relative to set.mm is largely due to also handling the ( # ` S ) < L
case (in which the prefix word ends up being empty, which also satisfies
the conclusion).