Add edgvalg to iset.mm (#5146)
* Add edgvalg to iset.mm

This is edgval from set.mm with an additional set existence condition.
The proof is via fvmptd3 and ifexd .

* Add iedgedgg to iset.mm

This is iedgedg from set.mm with a set existence condition added.  The
proof is the set.mm proof with small changes.

* add edgiedgbg to iset.mm

This is edgiedgb from set.mm with an additional set existence condition
added.  The proof is the set.mm proof with small changes.

* add edg0iedg0g to iset.mm

This is edg0iedg0 from set.mm with a set existence condition added.  The
proof is the set.mm proof with small changes.