Intuitionize prefix of words (#5148)
* Add prefixes of a word to set.mm

This is the syntax "prefix" and df-pfx .  Copied without change from
set.mm.

* Add pfxnndmnd to mmil.html

* Add pfxval to iset.mm

Stated as in set.mm.  The proof is the set.mm proof but a fair number of
set existence steps are needed to replace ovexd in the set.mm proof.

* add pfx00g to iset.mm

This is pfx00 from set.mm with a set existence condition added.  The
proof is via pfxval and swrd00g .

* Add pfx0g to iset.mm

This is pfx0 from set.mm with a condition added.  The proof is based on
a portion of the set.mm proof.

* Add pfxval0 to mmil.html

* add pfxclg to iset.mm

This is pfxcl with a condition added.  The proof is based on a portion
of the set.mm proof.

* copy pfxmpt from set.mm to iset.mm

* copy pfxres from set.mm to iset.mm

* copy pfxf from set.mm to iset.mm

* copy pfxfn from set.mm to iset.mm

* copy pfxfv from set.mm to iset.mm

* Add pfxlen to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* copy pfxid from set.mm to iset.mm

* copy pfxrn from set.mm to iset.mm

* copy pfxn0 from set.mm to iset.mm

* copy pfxnd from set.mm to iset.mm

The only change is a small adjustment to the comment.

* add pfxnd0 to mmil.html

unused in set.mm

* Add pfxwrdsymbg to iset.mm

This is pfxwrdsymb from set.mm with an additional condition.  The proof
is via pfxval and swrdwrdsymbg in the L <_ ( # ` S ) case (this part is
similar to the set.mm proof), but the additional length of the proof
relative to set.mm is largely due to also handling the ( # ` S ) < L
case (in which the prefix word ends up being empty, which also satisfies
the conclusion).

* copy addlenrevpfx from set.mm to iset.mm

* copy addlenpfx from set.mm to iset.mm

* copy pfxfv0 from set.mm to iset.mm

* copy pfxtrcfv from set.mm to iset.mm

* copy pfxtrcfv0 from set.mm to iset.mm

* Add pfxfvlsw to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* Add pfxeq to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a few small
changes.

* copy pfxtrcfvl from set.mm to iset.mm

* copy pfxsuffeqwrdeq from set.mm to iset.mm

* Add lswex to iset.mm

This is a set existence theorem which would be handled by fvex in
set.mm.  The proof is via lsw0g or lswcl for empty or nonempty
words, respectively.

* Add pfxsuff1eqwrdeq to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in
several places.

* copy invdisjrab from set.mm to iset.mm

the only change is removing a $j usage which does not hold in iset.mm.

* copy disjwrdpfx from set.mm to iset.mm

* Add ccatpfx to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes.

* copy fnssresd from set.mm to iset.mm

* copy pfxccat1 from set.mm to iset.mm

* copy pfx1 from set.mm to iset.mm

* Fix pfxres comment

This is in both set.mm and iset.mm. The new wording is
"Value of the prefix extractor as the restriction of a word" as
suggested in a pull request comment.

* remove addlenrevpfx from iset.mm and set.mm

Revise the one proof which had used it - lenrevpfxcctswrd in
set.mm - to use addlenpfx and addcomd .